feat(fiscal): extract fiscal module into packages/fiscal
Move fiscal code from apps/web/src/lib/fiscal/ to packages/fiscal/ as
standalone workspace package @finopenpos/fiscal with its own deps
(xml-crypto, fast-xml-parser), barrel export, and test suite. App now
imports from the workspace package instead of relative paths.
