+++ apps/web/package.json
@@ -15,6 +15,7 @@
   },
   "dependencies": {
     "@electric-sql/pglite": "^0.3.15",
+    "@finopenpos/fiscal": "workspace:*",
     "@faker-js/faker": "^10.3.0",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
@@ -40,7 +41,6 @@
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
     "drizzle-orm": "^0.45.1",
-    "fast-xml-parser": "^5.4.2",
     "lucide-react": "catalog:",
     "next": "catalog:",
     "next-intl": "^4.8.3",
@@ -52,7 +52,6 @@
     "tailwind-merge": "^2.4.0",
     "trpc-to-openapi": "^3.1.0",
     "tw-animate-css": "^1.4.0",
-    "xml-crypto": "^6.1.2",
     "zod": "catalog:"
   },
   "devDependencies": {
