+++ apps/web/src/lib/trpc/routers/fiscal-settings.ts
@@ -3,7 +3,7 @@ import { protectedProcedure, router } from "../init";
 import { db } from "@/lib/db";
 import { fiscalSettings } from "@/lib/db/schema";
 import { eq } from "drizzle-orm";
-import { getCertificateInfo } from "@/lib/fiscal/certificate";
+import { getCertificateInfo } from "@finopenpos/fiscal/certificate";
 import { checkSefazStatus } from "@/lib/invoice-service";
 
 export const fiscalSettingsRouter = router({
