+++ packages/fiscal/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@finopenpos/fiscal",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "default": "./src/index.ts"
+    },
+    "./*": {
+      "default": "./src/*.ts"
+    }
+  },
+  "scripts": {
+    "test": "bun test src/__tests__/",
+    "check-types": "tsc --noEmit"
+  },
+  "dependencies": {
+    "fast-xml-parser": "^5.4.2",
+    "xml-crypto": "^6.1.2"
+  },
+  "devDependencies": {
+    "@finopenpos/config": "workspace:*",
+    "@types/bun": "^1.3.9",
+    "typescript": "^5"
+  }
+}
