+++ packages/fiscal/src/__tests__/qrcode.test.ts
