+++ packages/fiscal/src/__tests__/xml-builder.test.ts
