+++ packages/fiscal/src/qrcode.ts
