+++ packages/fiscal/src/sefaz-event-types.ts
