+++ packages/fiscal/src/txt-structures.ts
