+++ packages/fiscal/src/valid-txt.ts
