+++ packages/fiscal/src/xml-builder.ts
