+++ packages/fiscal/src/xml-utils.ts
