+++ packages/fiscal/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@finopenpos/config/tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "rootDir": "./src"
+  },
+  "include": ["src/**/*.ts"],
+  "exclude": ["node_modules", "dist"]
+}
